requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.3
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
html5lib==1.1
idna==2.10
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
webencodings==0.5.1
whitenoise==5.2.0
